+++ CHANGELOG.rst
@@ -5,6 +5,11 @@ Version History
 Unreleased
 ----------
 
+* Add new (optional) parameters to ``ImportApi.import_files``,
+  ``BulkImportApi.bulk_import_upload_file`` and ``BulkImport.upload_file``. (#85)
+  The ``dtypes`` and ``converters`` parameters allow better control of the
+  import of CSV data (#83). This is modelled on the approach taken by pandas.
+
 v1.1.0 (2019-10-16)
 --------------------